rename focused-series.ts to nearby-series.ts, update imports
